Check QR code validation of the tickets code using python
The code works on checking the validations of the QR code by using the camera. If it is not from the images that were added to the system, it prints that it is invalid. If the QR code has been checked more than once, it will print invalid. So, basically, the code works on making sure that the scanned QR code on the tickets is from the database images and only checked once.